pyproject.toml
[tool.poetry]
name = "nonebot-plugin-escape-url"
version = "0.1.3.post1"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
readme = "README.md"
packages = [{ include = "nonebot_plugin_escape_url" }]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0"

[tool.poetry.group.dev.dependencies]
nonebot2 = { extras = ["httpx"], version = "^2.0.0rc4" }
nonebot-adapter-qqguild = "^0.2.2"
nonebot-adapter-onebot = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

